requirements.txt
args==0.1.0
asgiref==3.5.2
asyncio==3.4.3
boto3==1.24.31
botocore==1.27.31
certifi==2019.6.16
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
clint==0.5.1
coverage==4.0.3
coveralls==1.1
cryptography==37.0.2
curdling==0.4.0
defusedxml==0.7.1
distlib==0.1.2
dj-database-url==0.5.0
Django==4.0.4
django-appconf==1.0.5
django-countries==7.3.2
django-environ==0.9.0
django-excel==0.0.10
django-extensions==3.2.1
django-heroku==0.3.1
django-imagekit==4.1.0
django-ipware==4.0.2
django-mathfilters==1.0.0
django-qr-code==3.0.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.12
funcsigs==0.4
geoip2==2.9.0
gunicorn==20.1.0
heroku==0.1.1
html2image==2.0.1
html2pdf==0.1
htmlwebshot==0.1.0
httpretty==0.8.14
idna==2.7
imgkit==1.2.2
jmespath==1.0.1
lml==0.1.0
m2r==0.2.1
maxminddb==2.2.0
mistune==0.8.4
mock==1.3.0
mysqlclient==2.1.1
oauthlib==3.2.0
paystackapi==2.1.0
pbr==1.8.1
pilkit==2.0
Pillow==9.3.0
pkginfo==1.4.2
psycopg2==2.9.5
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
s3transfer==0.6.0
segno==1.5.2
simplejson==3.8.2
six==1.10.0
sqlparse==0.4.2
sure==1.2.24
texttable==1.6.4
tqdm==4.25.0
twine==1.11.0
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.10
whitenoise==6.1.0
